fix pb linter and package json
