+++ Rick-Morty/.vscode/extensions.json
@@ -1,0 +1,3 @@
+{
+  "recommendations": ["johnsoncodehk.volar", "johnsoncodehk.vscode-typescript-vue-plugin"]
+}
