+++ Rick-Morty/src/components/Filters.vue
@@ -38,7 +38,6 @@
 
 <script>
 
-  import { mapGetters } from 'vuex'
 
 
   export default {
@@ -80,11 +79,8 @@
     },
 
     computed: {
-      ...mapGetters([
-        'characters',
-        'getCharDisplayed',
-        'getCountCharacters'
-      ])
+     
+      
     }
    }
