+++ Rick-Morty/src/views/Character.vue
