+++ Rick-Morty/src/views/Home.vue
@@ -9,6 +9,7 @@
 import ListOfCharacters from '@/components/ListOfCharacters.vue'
 
 
+
 </script> 
 
 <style scoped>
