--- new_project/README.md
@@ -1,33 +1,0 @@
-# RICK AND MORTY X JELLYSMACK
-
-## About
-
-Hi ! Here is a data viz project i made to have a look on Rick and Morty characters.
-
-The call url to API is : 'https://rickandmortyapi.com/api/character'. 
-
-Enjoy :-)
-
-## Project setup
-```
-npm install
-```
-
-### Compiles and hot-reloads for development
-```
-npm run serve
-```
-
-### Compiles and minifies for production
-```
-npm run build
-```
-
-### Lints and fixes files
-```
-npm run lint
-```
-
-### Customize configuration
-See [Configuration Reference](https://cli.vuejs.org/config/).
-
